crawler/requirements.txt
akshare
pandas
psycopg2-binary
SQLAlchemy
schedule
yfinance
akshare